requirements.txt
annotated-types==0.7.0
appnope==0.1.4
asttokens==3.0.0
cairocffi==1.7.1
CairoSVG==2.8.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
comm==0.2.3
contourpy==1.3.3
cryptography==45.0.6
cssselect2==0.8.0
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
et_xmlfile==2.0.0
executing==2.2.0
fonttools==4.59.0
ibm-cloud-sdk-core==3.24.2
ibm-platform-services==0.68.1
idna==3.10
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
matplotlib==3.10.5
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
parso==0.8.4
pbr==6.1.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pylatexenc==2.10
pyparsing==3.2.3
pyspnego==0.11.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.0.1
qiskit==1.4.3
qiskit-aer==0.17.1
qiskit-ibm-runtime==0.41.0
qiskit-machine-learning==0.8.2
rdkit==2025.3.5
requests==2.32.4
requests_ntlm==1.3.0
rustworkx==0.16.0
SciencePlots==2.1.1
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
stevedore==5.4.1
symengine==0.13.0
sympy==1.14.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
webencodings==0.5.1
